Refactor `SchemaDumper` to make it possible to adapter specific customization
Refer https://github.com/rails/rails/pull/30337
