+++ lib/active_record/connection_adapters/oracle_enhanced_adapter.rb
@@ -36,7 +36,6 @@ require "active_record/connection_adapters/oracle_enhanced/connection"
 require "active_record/connection_adapters/oracle_enhanced/database_statements"
 require "active_record/connection_adapters/oracle_enhanced/schema_statements"
 require "active_record/connection_adapters/oracle_enhanced/schema_statements_ext"
-require "active_record/connection_adapters/oracle_enhanced/column_dumper"
 require "active_record/connection_adapters/oracle_enhanced/context_index"
 require "active_record/connection_adapters/oracle_enhanced/column"
 require "active_record/connection_adapters/oracle_enhanced/quoting"
@@ -163,7 +162,6 @@ module ActiveRecord
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::DatabaseStatements
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::SchemaStatements
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::SchemaStatementsExt
-      include ActiveRecord::ConnectionAdapters::OracleEnhanced::ColumnDumper
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::ContextIndex
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::Quoting
       include ActiveRecord::ConnectionAdapters::OracleEnhanced::DatabaseLimits
